Add functionality to remove file
